Get rid of DavDocumentsActor
