+++ app/src/main/kotlin/at/bitfire/davdroid/webdav/DavHttpClientBuilder.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
+ */
+
+package at.bitfire.davdroid.webdav
+
+import at.bitfire.davdroid.network.HttpClient
+import at.bitfire.davdroid.network.MemoryCookieStore
+import okhttp3.CookieJar
+import okhttp3.logging.HttpLoggingInterceptor
+import javax.inject.Inject
+import javax.inject.Provider
+
+class DavHttpClientBuilder @Inject constructor(
+    private val credentialsStore: CredentialsStore,
+    private val httpClientBuilder: Provider<HttpClient.Builder>
+) {
+
+    /**
+     * Creates an HTTP client that can be used to access resources in the given mount.
+     *
+     * @param mountId    ID of the mount to access
+     * @param logBody    whether to log the body of HTTP requests (disable for potentially large files)
+     */
+    fun build(mountId: Long, logBody: Boolean = true): HttpClient {
+        val builder = httpClientBuilder.get()
+            .loggerInterceptorLevel(if (logBody) HttpLoggingInterceptor.Level.BODY else HttpLoggingInterceptor.Level.HEADERS)
+            .setCookieStore(
+                cookieStores.getOrPut(mountId) { MemoryCookieStore() }
+            )
+
+        credentialsStore.getCredentials(mountId)?.let { credentials ->
+            builder.authenticate(host = null, getCredentials = { credentials })
+        }
+
+        return builder.build()
+    }
+
+
+    companion object {
+
+        // TODO global leak
+        private val cookieStores = mutableMapOf<Long, CookieJar>()
+
+    }
+
+}
