+++ app/src/main/kotlin/at/bitfire/davdroid/webdav/operation/QueryChildDocumentsOperation.kt
@@ -7,25 +7,44 @@ package at.bitfire.davdroid.webdav.operation
 import android.content.Context
 import android.provider.DocumentsContract.Document
 import android.provider.DocumentsContract.buildChildDocumentsUri
+import at.bitfire.dav4jvm.DavCollection
+import at.bitfire.dav4jvm.Response
+import at.bitfire.dav4jvm.property.webdav.CurrentUserPrivilegeSet
+import at.bitfire.dav4jvm.property.webdav.DisplayName
+import at.bitfire.dav4jvm.property.webdav.GetContentLength
+import at.bitfire.dav4jvm.property.webdav.GetContentType
+import at.bitfire.dav4jvm.property.webdav.GetETag
+import at.bitfire.dav4jvm.property.webdav.GetLastModified
+import at.bitfire.dav4jvm.property.webdav.QuotaAvailableBytes
+import at.bitfire.dav4jvm.property.webdav.QuotaUsedBytes
+import at.bitfire.dav4jvm.property.webdav.ResourceType
 import at.bitfire.davdroid.R
 import at.bitfire.davdroid.db.AppDatabase
+import at.bitfire.davdroid.db.WebDavDocument
 import at.bitfire.davdroid.db.WebDavDocumentDao
+import at.bitfire.davdroid.di.IoDispatcher
+import at.bitfire.davdroid.webdav.DavDocumentsProvider.Companion.DAV_FILE_FIELDS
+import at.bitfire.davdroid.webdav.DavHttpClientBuilder
 import at.bitfire.davdroid.webdav.DocumentSortByMapper
 import at.bitfire.davdroid.webdav.DocumentsCursor
 import dagger.Lazy
 import dagger.hilt.android.qualifiers.ApplicationContext
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.runInterruptible
 import java.io.FileNotFoundException
 import java.util.concurrent.ConcurrentHashMap
+import java.util.logging.Level
 import java.util.logging.Logger
 import javax.inject.Inject
 
 class QueryChildDocumentsOperation @Inject constructor(
-    private val actor: DavDocumentsActor,
     @ApplicationContext private val context: Context,
-    db: AppDatabase,
+    private val db: AppDatabase,
     private val documentSortByMapper: Lazy<DocumentSortByMapper>,
+    private val httpClientBuilder: DavHttpClientBuilder,
+    @IoDispatcher private val ioDispatcher: CoroutineDispatcher,
     private val logger: Logger
 ) {
 
@@ -65,7 +84,7 @@ class QueryChildDocumentsOperation @Inject constructor(
         // Dispatch worker querying for the children and keep track of it
         val running = runningQueryChildren.getOrPut(parentId) {
             externalScope.launch {
-                actor.queryChildren(parent)
+                queryChildren(parent)
                 // Once the query is done, set query as finished (not running)
                 runningQueryChildren[parentId] = false
                 // .. and notify - effectively calling this method again
@@ -94,6 +113,78 @@ class QueryChildDocumentsOperation @Inject constructor(
         }
 
         return result
+    }
+
+    /**
+     * Finds children of given parent [WebDavDocument]. After querying, it
+     * updates existing children, adds new ones or removes deleted ones.
+     *
+     * There must never be more than one running instance per [parent]!
+     *
+     * @param parent    folder to search for children
+     */
+    internal suspend fun queryChildren(parent: WebDavDocument) {
+        val oldChildren = documentDao.getChildren(parent.id).associateBy { it.name }.toMutableMap() // "name" of file/folder must be unique
+        val newChildrenList = hashMapOf<String, WebDavDocument>()
+
+        val parentUrl = parent.toHttpUrl(db)
+        httpClientBuilder.build(parent.mountId).use { client ->
+            val folder = DavCollection(client.okHttpClient, parentUrl)
+
+            try {
+                runInterruptible(ioDispatcher) {
+                    folder.propfind(1, *DAV_FILE_FIELDS) { response, relation ->
+                        logger.fine("$relation $response")
+
+                        val resource: WebDavDocument =
+                            when (relation) {
+                                Response.HrefRelation.SELF ->       // it's about the parent
+                                    parent
+
+                                Response.HrefRelation.MEMBER ->     // it's about a member
+                                    WebDavDocument(mountId = parent.mountId, parentId = parent.id, name = response.hrefName())
+
+                                else -> {
+                                    // we didn't request this; log a warning and ignore it
+                                    logger.warning("Ignoring unexpected $response $relation in $parentUrl")
+                                    return@propfind
+                                }
+                            }
+
+                        val updatedResource = resource.copy(
+                            isDirectory = response[ResourceType::class.java]?.types?.contains(ResourceType.COLLECTION)
+                                ?: resource.isDirectory,
+                            displayName = response[DisplayName::class.java]?.displayName,
+                            mimeType = response[GetContentType::class.java]?.type,
+                            eTag = response[GetETag::class.java]?.takeIf { !it.weak }?.let { resource.eTag },
+                            lastModified = response[GetLastModified::class.java]?.lastModified?.toEpochMilli(),
+                            size = response[GetContentLength::class.java]?.contentLength,
+                            mayBind = response[CurrentUserPrivilegeSet::class.java]?.mayBind,
+                            mayUnbind = response[CurrentUserPrivilegeSet::class.java]?.mayUnbind,
+                            mayWriteContent = response[CurrentUserPrivilegeSet::class.java]?.mayWriteContent,
+                            quotaAvailable = response[QuotaAvailableBytes::class.java]?.quotaAvailableBytes,
+                            quotaUsed = response[QuotaUsedBytes::class.java]?.quotaUsedBytes,
+                        )
+
+                        if (resource == parent)
+                            documentDao.update(updatedResource)
+                        else {
+                            documentDao.insertOrUpdate(updatedResource)
+                            newChildrenList[resource.name] = updatedResource
+                        }
+
+                        // remove resource from known child nodes, because not found on server
+                        oldChildren.remove(resource.name)
+                    }
+                }
+            } catch (e: Exception) {
+                logger.log(Level.WARNING, "Couldn't query children", e)
+            }
+        }
+
+        // Delete child nodes which were not rediscovered (deleted serverside)
+        for ((_, oldChild) in oldChildren)
+            documentDao.delete(oldChild)
     }
